UPDATE Refactor Dashboard into components & hooks
Split the large Dashboard implementation into smaller components and hooks for better modularity and testability. Added DashboardHeader, DashboardGrid and DashboardFooter components and replaced the big inline header/grid/footer markup. Extracted WebSocket and backend session logic into useDashboardWebSocket and useDashboardSession hooks (handles connection, stream fetching, buffering, reconnects, and Electron IPC). Exported SIGNAL_COLORS from the monitor module and adjusted monitor internals (canvas component rename, min size units, and color selection). Updated Dashboard to compose the new components/hooks, derive recording state from connection, and use a uid helper for monitor IDs.
